(PC-35367) refactor(offer): use interaction tag react node rather than just likes counter
