+++ src/features/offer/components/OfferTile/OfferTile.stories.tsx
@@ -2,6 +2,11 @@ import { NavigationContainer } from '@react-navigation/native'
 import { ComponentMeta, ComponentStory } from '@storybook/react'
 import React from 'react'
 import { View } from 'react-native'
+import styled from 'styled-components/native'
+
+import { Tag } from 'ui/components/Tag/Tag'
+import { ThumbUpFilled } from 'ui/svg/icons/ThumbUpFilled'
+import { getSpacing } from 'ui/theme'
 
 import { OfferTile } from './OfferTile'
 
@@ -18,6 +23,17 @@ const meta: ComponentMeta<typeof OfferTile> = {
 }
 
 export default meta
+
+const CustomThumbUp = styled(ThumbUpFilled).attrs(({ theme }) => ({
+  size: getSpacing(4),
+  color: theme.colors.primary,
+}))``
+
+const LikeTag = styled(Tag).attrs(() => ({
+  Icon: CustomThumbUp,
+}))(({ theme }) => ({
+  backgroundColor: theme.colors.white,
+}))
 
 const Template: ComponentStory<typeof OfferTile> = (props) => <OfferTile {...props} />
 
@@ -40,6 +56,6 @@ WithTags.args = {
   categoryLabel: 'Cinéma',
   width: 200,
   height: 300,
-  likes: 99,
+  interactionTag: <LikeTag label="100" />,
   offerLocation: { lat: 48.94374, lng: 2.48171 },
 }
