+++ src/features/offer/components/OfferTile/OfferTile.tsx
@@ -32,7 +32,7 @@ const UnmemoizedOfferTile = (props: OfferTileProps) => {
     apiRecoParams,
     index,
     artistName,
-    likes,
+    interactionTag,
     navigationMethod = NAVIGATION_METHOD.NAVIGATE,
     ...offer
   } = props
@@ -108,7 +108,7 @@ const UnmemoizedOfferTile = (props: OfferTileProps) => {
           categoryId={categoryId}
           thumbnailUrl={thumbUrl}
           distance={distanceFromOffer}
-          likes={likes}
+          interactionTag={interactionTag}
           name={name}
           date={date}
           price={price}
